split metadata.py into three, creating freedb.py and musicbrainz.py
